Added contributor count extraction through Selenium.
If a page fails to locate contributors using Jsoup, a chomedriver is launched to extract the data. Currently known issues: 1) Sometimes the ChromeDriver.get(...) method may throw an UnreachableBrowserException. This happens rarely and non-deterministically, but it can happen. The exact cause is not known, but for now, whenever chromedriver fails to launch, contributors are set to -1. A job can be created to re-crawl the contributors for these specific repositories. 2) After ~40 created chromedrivers, a IOException, stating that there are "Too many open files". I have yet to find a fix for this...