CrawlJob repository not updated for 0 results retrieved.
A problem I encountered when crawling the last 2 hours of created repos, is that
GitHub does not manage to update their DBs to account for these projects.
Retrieving the last 2 hours of newly created projects yields no results. A
temporary (and imperfect) workaround to this is not to update the interval when
there are no results retrieved.