Added distinct retry mechanism in case CrawlerService throws a 429 HttpStatusException
Sometimes GitHub may freak out and return a 429 when attempting to mine the data from the
pages. A 429 was previously unhandled, and the repo, even though existing, would not be
saved. Not only that, but any subsequent repo mined in the next few seconds would also not
be saved. Since this issue occurs infrequently, I have opted to add a cyclic retry after
1 minute of waiting. In cases encountered so far, waiting for 1 minute is more than enough,
and the data is mined as early as the second attempt.