Added basic 5XX error handling
In case of an error being returned from GitHub, the crawler waits for a certain amount
of time (I set it to 1 hour), and then attempts to perform the request again. I'm not
too concerned with spin-waits, as GitHub is usually down for just about one to two hours
according to their API status statistics. To preserve cohesion and avoid repeating certain
tasks, I added the new CrawlProjectsJob.retrieveRemainingRepos method. While
CrawlProjectsJob.retrieveRepos is only concerned with the first 100 results of an interval,
the newly defined method retrieves and saves the remaining results.