Added Ranges utility
This marks the beginning of removing the poorly implemented Interval
interface. Said class (and concrete children) will be completely
removed in favor of Guava's Range interface. While far better in
terms of implementation, the class lacks several features which the
project demands. Given that the class is final, we use a static
utility class, Ranges, which will handle the building, splitting
and conversion to string.