Exposed custom headers to CORS configuration
Although working locally, the `Links` and `Download`
headers would normally not be sent to the FE due to CORS
rules regarding exposed headers. This is no longer the
case, and any client making calls to the API will be
able to receive these custom + all the ready-made headers
(i.e. `Content-Type`, `Date`, ...)