Improved handling for "null" stats of repositories
null rows used to occur due to problems with css
selectors that were used by jsoup/selenium to extract
data. Given that these selectors did not work when the
information was not on the page, and that this
information, was not on the page when its count was 0
(i.e. no contributor section, because there were no
project contributors) it is safe to assume that it was
equal to 0. Although it will take time to correct this
in the database, for the time being we simply replace
nulls in numeric data with a 0.