Improved link builder
Fixes #27

The bug was caused by the fact that string parameter inputs such as "C++"
would not be encoded in the other generated links. Simply encoding such params
manually would not be enough, as they would then get double-encoded by the link
builder. By changing the link building strategy we resolve such problem. Note
that this is not the most optimal solution, as we hard-code the endpoints
within the method itself. When I have more time, I will go back and further
simplify the link construction approach.