Optimizing `CleanUpProjectsJob`
Now uses id-based iteration instead of loading all project names which
will be checked into memory. Both approaches have their strengths and
weaknesses. The former executes only one query, but leads all the
results into memory. The latter does `n` total database reads, where
`n` is the number of rows in the `git_repo` table.

Q: Why not use streams?
A: Because the deletions are not immediately visible in the DB, the
   cleaning results would only be noticed after a single pass over the
   database. This could take several days, not exactly optimal in terms
   of serving up-to-date data.