Added a header attachment interceptor
This one will be used to set default request headers at a global level.
The `Accept` HTTP header is the first thing that comes to mind. In its
current state, it only allows a single key and value to be specified,
although the implementation can perhaps be extended later to use a
`Multimap` or something.