Creating converters from `JsonObject` to `RateLimit` and `GitCommit`
Given that the information used to construct these will be received from
the GitHub API, we should prefer the usage of a conversion service over
private methods to make the conversion process reusable. The extraction
logic is based on the one already present in `GitHubApiService` and
`GitHubCredentialUtil`. The two converters have also been registered.