Introducing `Endpoint`
Meant to serve as a replacement enum for `Endpoints`. Apart from a
correct singular name, the replacement offers URI Templates instead of
Strings, meaning that a URL object can be obtained from the enum, which
is then meant to be used in requests.

Note that `{owner}/{name}` had to be used instead of just `{name}` due
to HATEOAS encoding the `/` character if present in a parameter, which
as you might have guessed all repository names have.