`CleanUpProjectsJob` now uses randomized result scroll
Although we are mixing in the lower-level Hibernate with purely JPA
server code, I have found that the usage of `ScrollableResults` spawned
by a `StatelessSession` is a far superior solution. It not only performs
better, but we are also able to do things like randomize the order of
repositories during cleanup. At the same time, by explicitly managing
transactions, the cleaning deletions take immediate effect in the DB.