Added `logback` configuration for production
Segregates the output into 4 categories:
- `CONSOLE`: Prints all statements above and including the INFO level;
- `FILE-GITHUB`: Writes all interactions with the GitHub API to a file;
- `FILE-CRAWLER`: Writes all crawling details to a file;
- `FILE-CLEANER`: Writes all cleaning details to a file.

Note that this configuration is reserved for server deployment and is
not used by default. This is because the console-centric setup is better
for local development, whereas the segregated setup is better for
server-side deployment. In short the server console should be kept clean
while the files should contain all the details.