Changed single-result getter signature for services
They now return an object directly instead of just propagating the
`Optional` obtained from the DB. This approach has pros and cons
associated with it. The pros are that exceptions can be easily mapped
to the REST controllers to return specific responses when they are
thrown. The downside is that other code depending on these methods now
has to employ exception handling.