Added repeatable migration for `label` and `topic` ranking
These two "cached" views, which are just standalone read-only tables,
store the rankings for labels and topics. Both have 3 triggers that
prevent INSERT, UPDATE and DELETE from being run on them. Lastly, the
inclusion of the timestamp comment ensures that the repeatable migration
will be run prior to each server startup. In essence, it servers to
"refresh" the rankings of the two tables.