Repos that have (or could) not be analyzed are now iterated randomly
This is to ensure that we do not waste time always looking at the
constantly failing analysis attempts that will accumulate on top over
time. But this will not solve the issue fully. Later, we will be left
with a small starting pool of repositories that could not be analyzed.
I think we should still update the `cloned` date even in certain
unsuccessful cases.