Separated `Language` statistics into their own table and relation
The inner class `Statistics` of `Language` now houses the total count
of mined and analyzed repositories for each language. The data itself
is stored in a separate table, which is refreshed every 5 minutes by
a database event. The API endpoint for the statistics was changed to
include both counts, while the chart in the website was also changed
to display both stats, and their total counterparts.