Reverting the revert of feature: cleaning during analysis
The root cause was that the delete transaction was not set with the
`REQUIRES_NEW` propagation. During the cleanup job, a new one is
created regardless, but this change allows us to also use it during
the analysis!

(cherry picked from commit a39594d489cb4e1886edd591c1ec8bfb81ac7069)