Restoring `CleanUpProjectsJob` to use streaming like it did before
The only difference now is that the process is more optimized. It will
be triggered every monday, streaming through all repositories that have
not been pinged for more than 5 weeks. At the same time, it performs
the old ls-remote check, with a cURL fallback. If the repository can
not be reached by either means, it is deleted. Otherwise, we update the
ping value.