Added new archetype: `Connector`
While it's purpose is documented, I want to highlight _why_ this was
introduced. The `CommandLineInterfaceDependencyBean`, while faster, was
not ideal at identifying a single point of failure. I also prefer this
solution, because the dependency on the CLI program is immediately
specified in the component implementation. This puts more responsibility
on the programmer to explicitly name what he truly needs.