Add `pipefail` option to alpine shell in website build
Added to address the following `hadolint` warning:

Set the SHELL option -o pipefail before RUN with a pipe in it.
If you are using /bin/sh in an alpine image or if your shell is
symlinked to busybox then consider explicitly setting your SHELL
to /bin/ash, or disable this check.

Solution reference:

https://github.com/hadolint/hadolint/wiki/DL4006