Add `SimpleErrorController`
Intended to override `BasicErrorController`. This controller only
returns the status code, as opposed to the HTML which is the default
error controller behaviour. To prevent confusion that may stem from
manually navigating to `/api/error`, the route is instructed to return
a 404 not found. In case of other errors, the appropriate status code
will be returned instead.