Overhaul the `download` endpoint
This contains too many changes to be split into smaller, consecutive
commits, so I will group them all into a larger commit. Why was this
change even made in the first place? Originally, I intended to add
the new `ExportFormat` and call it a day, but I noticed that there
were some memory allocation issues in the download that were caused
by:

1) Not using a `@PersistenceContext` entity manager
2) Not clearing the entity manager cache on a regular basis

With both changes, the memory issues have been addressed. After that,
my goal was only to improve the readability of the overall code. The
new code did see the removal of the `exportFormats` bean, as well as
changes to the `DownloadLinkBuilder`. Another change is that CSV
column values will now always be quoted.