Replace `PersistenceContextInvalidator` with `PeriodicCallback`
The advantage of this new class is that it accepts an arbitrary number
of `Runnable` instances to execute when a specified counter condition
is met. This makes the code more reusable overall.

The reason this was done was to allow periodic flushes of the
`JsonGenerator`. Although I'm not 100% sure if this improves the
performance, we will revert it's flushing if it is determined to
be of no use.

This change was also the reason we introduced the `CheckedRunnable`
and its `IOException`-specialized counterpart.