Use default CSV serialization for lists
I have no idea why we did this in the first place. I guess whoever
implemented this never bothered to check if the lists can be serialized
in CSV. By using the default serialization we also inadvertently save up
on space, as empty arrays will not be printed as `[]` to the CSV, but
rather a `null` in the form of an empty column value.