Further compress PWA icons
Although their resolution remains the same, the color depth was reduced
from a 24-bit to an 8-bit palette. Furthermore, all profiles comments
and PNG chunks were removed. Lastly, a small issue was fixed where some
of the higher-resolution icons were displaying a thin black bar on the
right side.This was all achieved with the help of `imagemagick`.