Filter all languages that have no entries mined in the chart
This is useful in cases when we add a lot of new languages in bulk.
We could have also modified the service method or REST API endpoint,
but I have no idea if we might end up needing the original behaviour
for something else. Ergo, the filtering is left up to the presentation
layer.