`body` now has `.modal-open` class on by default
This class is removed by JavaScript on load. The idea being that since
the `noscript` tag shows a modal, the class should also be present on
the body. I'm not quite sure why BS adds this class to the body whenever
you open modals, but it might have something to do with global CSS.