Hide `body` overflow when `noscript` modal is shown
Typically, when a modal is show the underlying JavaScript hides the
scrollbar (see line 114 of `modal.js` in their repository) for the
`body`. Since we can not use JavaScript, we emulate the same behaviour
through one of its utilities: we add the `.overflow-hidden` class to the
`body`, removing it via JS only once the page loads.