Integrate JGit into `CleanUpProjectsJob`
Still not 100% ready, should have more robust exception handling to
detect exact failures. `TransportException` is too broad, and may even
encapsulate cases where the check should be undefined as opposed to a
failure. Then again, this should not be too big of an issue, since we
check with HTTP in the event of Git failures.